renamed file with prompts
